+++ CHANGELOG.md
@@ -7,12 +7,16 @@ The format is based on the [KeepAChangeLog] project.
 
 ## Unreleased
 
+### Fixed
+- [#602] Fixed uncaught error on unpacking of message
+
 ### Removed
 - [#671] Removed deprecated request/response_cls kwargs from Provider/Client methods
 - [#677] Removed more deprecated code
 
 [#671]: https://github.com/OpenIDC/pyoidc/pull/671
 [#677]: https://github.com/OpenIDC/pyoidc/pull/677
+[#602]: https://github.com/OpenIDC/pyoidc/issues/602
 
 ## 1.0.1 [2019-06-30]